ntpMerlin compatibility

+++ spdmerlin.sh
@@ -567,10 +567,12 @@ Menu_Uninstall(){
 	done
 	Shortcut_spdMerlin delete
 	opkg remove --autoremove python
-	opkg remove --autoremove rrdtool
-	umount /www/require/modules/menuTree.js 2>/dev/null
 	umount /www/Advanced_Feedback.asp 2>/dev/null
-	rm -f "/jffs/scripts/custom_menuTree.js" 2>/dev/null
+	if [ ! -f "/jffs/scripts/ntpmerlin" ]; then
+		opkg remove --autoremove rrdtool
+		umount /www/require/modules/menuTree.js 2>/dev/null
+		rm -f "/jffs/scripts/custom_menuTree.js" 2>/dev/null
+	fi
 	rm -f "/jffs/scripts/custom_state.js" 2>/dev/null
 	rm -f "/jffs/scripts/spdstats_www.asp" 2>/dev/null
 	rm -f "/jffs/scripts/spdcli.py" 2>/dev/null
